+++ modules/frontend/frontend_test.go
@@ -38,7 +38,7 @@ func TestFrontendTagSearchRequiresOrgID(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, next, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, next, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	require.NoError(t, err)
 
 	req := httptest.NewRequest("GET", "/", nil)
@@ -74,7 +74,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend query shards should be between 2 and 100000 (both inclusive)")
 	assert.Nil(t, f)
 
@@ -91,7 +91,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend query shards should be between 2 and 100000 (both inclusive)")
 	assert.Nil(t, f)
 
@@ -108,7 +108,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend search concurrent requests should be greater than 0")
 	assert.Nil(t, f)
 
@@ -125,7 +125,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend search target bytes per request should be greater than 0")
 	assert.Nil(t, f)
 
@@ -144,7 +144,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "query backend after should be less than or equal to query ingester until")
 	assert.Nil(t, f)
 
@@ -160,7 +160,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend metrics concurrent requests should be greater than 0")
 	assert.Nil(t, f)
 
@@ -183,7 +183,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend metrics target bytes per request should be greater than 0")
 	assert.Nil(t, f)
 
@@ -207,7 +207,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "frontend metrics interval should be greater than 0")
 	assert.Nil(t, f)
 
@@ -231,7 +231,7 @@ func TestFrontendBadConfigFails(t *testing.T) {
 			},
 			SLO: testSLOcfg,
 		},
-	}, nil, nil, nil, nil, "", log.NewNopLogger(), nil)
+	}, nil, nil, nil, nil, "", fakeHTTPAuthMiddleware, log.NewNopLogger(), nil)
 	assert.EqualError(t, err, "most recent shards must be greater than 0")
 	assert.Nil(t, f)
 }